--- src/charts/chart-grid/ChartGrid.css
@@ -1,34 +1,0 @@
-@import '../../materials/colors';
-@import '../../materials/layout';
-
-.ax-chart-grid {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  height: 100%;
-}
-
-.ax-chart-grid__lines {
-  display: flex;
-  flex: 1 0 100%;
-  max-height: 100%;
-}
-
-.ax-chart-grid__lines-svg {
-  display: block;
-  flex: 1 1 0%;
-  width: 100%;
-}
-
-.ax-chart-grid__line {
-  stroke: var(--color-context-border-subtle);
-  vector-effect: non-scaling-stroke;
-}
-
-@media (--screen-less-than-small) {
-  .ax-chart-grid__line-svg:nth-child(even) {
-    & .ax-chart-grid__line {
-      display: none;
-    }
-  }
-}
