--- src/charts/chart-grid/ChartGrid.test.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import renderer from 'react-test-renderer';
-import { ChartGrid } from 'bw-axiom';
-
-function getComponent() {
-  return renderer.create(
-    <ChartGrid count={ 11 } />
-  );
-}
-
-describe('ChartGrid', () => {
-  it('renders with defaultProps', () => {
-    const component = getComponent();
-    const tree = component.toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-});
