+++ src/charts/chart-table/ChartTableGrid.js
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
-import { Base, ChartGrid } from 'bw-axiom';
+import { Base } from 'bw-axiom';
 import './ChartTableGrid.css';
 
 export default class ChartTableGrid extends Component {
@@ -20,6 +20,7 @@ export default class ChartTableGrid extends Component {
   render() {
     const { labelColumnWidth } = this.context;
     const { children, count, ...rest } = this.props;
+    const increments = Array.apply(null, { length: count }).map((_, i) => i * 10);
 
     return (
       <Base { ...rest } className="ax-chart-table__grid-container">
@@ -28,7 +29,26 @@ export default class ChartTableGrid extends Component {
         <div
             className="ax-chart-table__grid"
             style={ { left: labelColumnWidth } }>
-          <ChartGrid count={ count } />
+          <div className="ax-chart-table__grid-lines">
+            <svg className="ax-chart-table__grid-lines-svg">
+              { increments.map((x) =>
+                <svg
+                    className="ax-chart-table__grid-line-svg"
+                    key={ x }
+                    preserveAspectRatio="none"
+                    viewBox="0 0 1 1"
+                    width="1"
+                    x={ `${x}%` }>
+                  <path
+                      className="ax-chart-table__grid-line"
+                      d="M 0 0 V 1"
+                      shapeRendering="crispEdges"
+                      strokeDasharray="1, 1"
+                      strokeWidth="1" />
+                </svg>
+              ) }
+            </svg>
+          </div>
         </div>
       </Base>
     );
