+++ src/charts/chart-table/example/collapsible.js
@@ -1,34 +1,72 @@
 import React, { Component } from 'react';
 import { Example, Snippet } from 'style-guide';
 import {
+  Button,
+  ButtonGroup,
   ChartKey,
   ChartKeyItem,
   ChartTable,
   ChartTableAxis,
   ChartTableGrid,
   ChartTableKey,
-  ChartTableRows,
-  ChartTableRow,
   ChartTableLabel,
+  ChartTableRow,
+  ChartTableRows,
   ChartTableVisual,
-  DataPoints,
   DataPoint,
+  DataPoints,
   Dot,
   DotPlot,
   DotPlotLine,
 } from 'bw-axiom';
-import { labels, data } from './data';
+import { labels, data } from '../../dot-plot/example/data';
 
-export default class DotPlotExample extends Component {
+export default class ChartTableExample extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      rowData: data.slice(0, 7),
+    };
+  }
+
+  addRow() {
+    this.setState(({ rowData }) => ({
+      rowData: data.slice(0, rowData.length + 1),
+    }));
+  }
+
+  removeRow() {
+    this.setState(({ rowData }) => ({
+      rowData: data.slice(0, rowData.length - 1),
+    }));
+  }
+
   render() {
+    const { rowData } = this.state;
+
     return (
-      <Example name="DotPlot inside ChartTable">
+      <Example name="Collapsible ChartTable">
+        <ButtonGroup>
+          <Button
+              disabled={ rowData.length === data.length }
+              onClick={ () => this.addRow() }>
+            Add a row of data
+          </Button>
+          <Button
+              disabled={ rowData.length === 0 }
+              onClick={ () => this.removeRow() }>
+            Remove a row of data
+          </Button>
+        </ButtonGroup>
+
         <Snippet>
           <ChartTable
+              collapsedVisibleRowCount={ 6 }
+              expandButtonSuffix="Categories"
               labelColumnWidth="11rem">
-            <ChartTableGrid>
+            <ChartTableGrid snippetSkip={ true }>
               <ChartTableRows>
-                { data.map(({ label, data }, i) =>
+                { rowData.map(({ label, data }, i) =>
                   <ChartTableRow key={ label } snippetSkip={ i > 0 }>
                     <ChartTableLabel>{ label }</ChartTableLabel>
                     <ChartTableVisual>
@@ -45,8 +83,8 @@ export default class DotPlotExample extends Component {
                 ) }
               </ChartTableRows>
             </ChartTableGrid>
-            <ChartTableAxis title="% of each something" />
-            <ChartTableKey>
+            <ChartTableAxis  snippetSkip={ true } title="% of each something" />
+            <ChartTableKey snippetSkip={ true }>
               <ChartKey>
                 { labels.map(({ name, color }) =>
                   <ChartKeyItem key={ name } label={ name }>
