+++ src/charts/chart-table/example/index.js
@@ -1,0 +1,5 @@
+module.exports = [
+  require('../../bars/example/chart-table').default,
+  require('../../dot-plot/example/chart-table').default,
+  require('./collapsible').default,
+];
