+++ src/charts/dot-plot/example/chart-table.js
@@ -1,72 +1,34 @@
 import React, { Component } from 'react';
 import { Example, Snippet } from 'style-guide';
 import {
-  Button,
-  ButtonGroup,
   ChartKey,
   ChartKeyItem,
   ChartTable,
   ChartTableAxis,
   ChartTableGrid,
   ChartTableKey,
-  ChartTableRows,
-  ChartTableRow,
   ChartTableLabel,
+  ChartTableRow,
+  ChartTableRows,
   ChartTableVisual,
-  DataPoints,
   DataPoint,
+  DataPoints,
   Dot,
   DotPlot,
   DotPlotLine,
 } from 'bw-axiom';
 import { labels, data } from './data';
 
-export default class DotPlotExample extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      rowData: data.slice(0, 7),
-    };
-  }
-
-  addRow() {
-    this.setState(({ rowData }) => ({
-      rowData: data.slice(0, rowData.length + 1),
-    }));
-  }
-
-  removeRow() {
-    this.setState(({ rowData }) => ({
-      rowData: data.slice(0, rowData.length - 1),
-    }));
-  }
-
+export default class ChartTableExample extends Component {
   render() {
-    const { rowData } = this.state;
-
     return (
-      <Example name="DotPlot inside collapsible ChartTable">
-        <ButtonGroup>
-          <Button
-              disabled={ rowData.length === data.length }
-              onClick={ () => this.addRow() }>
-            Add a row of data
-          </Button>
-          <Button
-              disabled={ rowData.length === 0 }
-              onClick={ () => this.removeRow() }>
-            Remove a row of data
-          </Button>
-        </ButtonGroup>
-
+      <Example name="ChartTable with DotPlots">
         <Snippet>
           <ChartTable
-              collapsedVisibleRowCount={ 6 }
-              expandButtonSuffix="Categories"
               labelColumnWidth="11rem">
-            <ChartTableGrid snippetSkip={ true }>
+            <ChartTableGrid>
               <ChartTableRows>
-                { rowData.map(({ label, data }, i) =>
+                { data.map(({ label, data }, i) =>
                   <ChartTableRow key={ label } snippetSkip={ i > 0 }>
                     <ChartTableLabel>{ label }</ChartTableLabel>
                     <ChartTableVisual>
@@ -83,8 +45,8 @@ export default class DotPlotExample extends Component {
                 ) }
               </ChartTableRows>
             </ChartTableGrid>
-            <ChartTableAxis  snippetSkip={ true } title="% of each something" />
-            <ChartTableKey snippetSkip={ true }>
+            <ChartTableAxis title="% of each something" />
+            <ChartTableKey>
               <ChartKey>
                 { labels.map(({ name, color }) =>
                   <ChartKeyItem key={ name } label={ name }>
