+++ src/charts/dot-plot/example/index.js
@@ -1,5 +1,3 @@
 module.exports = [
-  require('./dot-plot').default,
-  require('./dot-plot-table').default,
-  require('./dot-plot-table-collapsible').default,
+  require('./chart-table').default,
 ];
