+++ src/index.js
@@ -101,9 +101,6 @@ export { default as Weak } from './components/typography/Weak';
 export { default as Bar } from './charts/bar-chart/Bar';
 export { default as BarChart } from './charts/bar-chart/BarChart';
 export { default as BarGroup } from './charts/bar-chart/BarGroup';
-export { default as ChartAxisLabels } from './charts/chart-axis/ChartAxisLabels';
-export { default as ChartAxisTitle } from './charts/chart-axis/ChartAxisTitle';
-export { default as ChartGrid } from './charts/chart-grid/ChartGrid';
 export { default as Bars } from './charts/bars/Bars';
 export { default as Bar } from './charts/bars/Bar';
 export { default as ChartHeading } from './charts/chart-heading/ChartHeading';
